Stream d'électricité ajouté Un bloc peut maintenant accueillir plusieurs stream en simultané, la propriété GetCurrentTypes retourne une liste de types de stream

+++ trunk/Unity project/Assets/Resources/Scripts/Sources/SourceInfo.cs
@@ -4,8 +4,8 @@ using System.Collections;
 public class SourceInfo : MonoBehaviour {
 
 	public Source.SourceType type;
-	public int generate;
-	public int duration;
+	public ushort generate;
+	public ushort duration;
 	public Mesh mesh;
 	public Material material;
